+++ docs/changelog.md
@@ -80,6 +80,10 @@
 
 ## 更新说明详情
 
+1.17.55 fix: fix stock_zh_growth_comparison_em interface
+
+	1. 修复 stock_zh_growth_comparison_em 接口
+
 1.17.54 fix: fix fund_announcement_dividend_em interface
 
 	1. 修复 fund_announcement_dividend_em 接口
@@ -5812,6 +5816,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.17.55 fix: fix stock_zh_growth_comparison_em interface
 
 1.17.54 fix: fix fund_announcement_dividend_em interface